Attempted resolve of two issues:
1) Searching repositories
When attempting to change the search term, the previous term is still passed through and the result is a concatenation of the two.

2) Viewing issues
When selecting a particular repository result in order to view it's issues, the repo 'full_name' is being passed through bu the API call is being called twice: the first time with the correct 'full_name' value, and the second with an 'undefined' 'full_name'. This second call is interrupting the initial API call and therefore failing. When removing the suspected reason for the failure, the API call does not get hit at all.